Rename remaining classes/interface
